Store sequences on each executor.

+++ spark-impl/src/main/scala/one/off_by/sequence/mining/gsp/Model.scala
@@ -38,5 +38,5 @@ case class Pattern[ItemType](
 object Domain {
   type Percent = Double
   type Support = Percent
-  type SupportCount = Long
+  type SupportCount = Int
 }
